326. Power of Three
